Modulo_Rekognition_EPP/requirements.txt
boto3
botocore
python-dotenv
opencv-python